[#6] refactor: atoms component import 경로 전환
