+++ frontend/src/components/atoms/Heading/index.ts
@@ -1,1 +1,1 @@
-export { Heading1, Heading2, Heading3 } from './Heading';
+export { default } from './Heading';
