+++ frontend/src/components/atoms/index.ts
@@ -1,2 +1,3 @@
-export * as HeaderLink from './HeaderLink';
-export * as HeaderButton from './HeaderButton';
+export { default as HeaderLink } from './HeaderLink';
+export { default as HeaderButton } from './HeaderButton';
+export { default as Heading } from './Heading';
